Rename loaderSpinner
