--- dashboard/src/components/LoaderSpinner/index.ts
@@ -1,3 +1,0 @@
-import LoaderSpinner from "./LoaderSpinner";
-
-export default LoaderSpinner;
