+++ dashboard/src/components/LoadingSpinner/LoadingSpinner.scss
